Add Windows support for the Python API (#1811)
* Call `string()` to convert a path to a string.

* Use `$<CONFIG>` instead of `${CMAKE_BUILD_TYPE}` in places easy to replace.

* Use the export header macros to export the `LOG_***` functions.

* Don't pass `-Wno-deprecated-declarations` in MSVC.

* Fix an MSVC overflow warning that turns into an error.

* Don't link to the `tiledbsoma` target in the unit test.

This target corresponds to the dynamic libtiledbsoma library, but we already link to it through `TILEDB_SOMA_OBJECTS`.

* Run Python CI on Windows.

* Small fixes in the Python CI workflows.

* Remove an outdated comment.

* Fix including `tiledbsoma_export.h`.

* Fix include directories.

* Add missing headers.

* Fix the Python minimal CI workflow.

It now runs on `pull_request` instead of `push`, covering PRs from forks (added `workflow_dispatch` for those that might still want to run it without opening a PR), and adds a Windows matrix job (only for python 3.10 to reduce the jobs of the workflow).
Also fixed two apparent typos.

* Add a Powershell script to build libtiledbsoma.

* Fix the build script.

* Fix symlink path resolution.

* Reformat.

* Import the `tiledb` library on Windows only.

* Add Windows support in the Python code.

* Install the runtime dlls of `tiledbsoma`.

* Fix the lint job on Windows.

* Run more steps on bash.

* Handle both `RuntimeError` and `TileDBError` in `test_types_read_errors`.

* Fix setting the `soma_object_type` metadata.

* Use `||` instead of `or`.

MSVC does not support it.

* Mention the actual type of `SOMA_JOINID` in case of mismatch.

* Force creating the `SOMA_JOINID` column with type int64.

* Ensure the coordinates are of type int64.

And avoid allocating temporary arrays in some cases.

* Explicitly set timestamps in the metadata tests.

* Add a delay of ten milliseconds to fix race conditions.

* Delete apis/python/src/tiledbsoma/tiledbsoma.dll

+++ libtiledbsoma/src/soma/column_buffer.h
@@ -298,7 +298,7 @@ class ColumnBuffer {
             std::string s(enums_[i]);
             enum_str_ += s;
             enum_offsets_[i] = cumlen;
-            cumlen += s.length();
+            cumlen += static_cast<uint32_t>(s.length());
         }
         enum_offsets_[n_vec] = cumlen;
     }
